requirements-dev.txt
-r requirements.txt
ipython[notebook]
Pillow
model-mommy==1.2.6
pip
whitenoise
django-test-without-migrations==0.4